done manager user

+++ MobileApp/movie_admin/lib/ui/users/manager_users_bloc.dart
@@ -1,26 +1,26 @@
 import 'package:disposebag/disposebag.dart';
 import 'package:flutter_bloc_pattern/flutter_bloc_pattern.dart';
 import 'package:meta/meta.dart';
-import 'manage_user_state.dart';
 import 'package:rxdart/rxdart.dart';
 
 import '../../domain/model/user.dart';
 import '../../domain/repository/manager_repository.dart';
 import '../../utils/utils.dart';
+import 'manage_user_state.dart';
 
 class ManagerUsersBloc extends DisposeCallbackBaseBloc {
   /// Input functions
   final Function1<int, void> loadUsers;
-  final Function1<User, void> removeUser;
+  final Function1<MapEntry<User, DestroyUserType>, void> destroyUser;
 
   /// Streams
   final Stream<ManageUserState> renderListStream$;
-  final Stream<List<String>> renderItemRemove$;
+  final Stream<Map<String, DestroyUserType>> renderItemRemove$;
 
   ManagerUsersBloc._({
     @required Function0<void> dispose,
     @required this.loadUsers,
-    @required this.removeUser,
+    @required this.destroyUser,
     @required this.renderListStream$,
     @required this.renderItemRemove$,
   }) : super(dispose);
@@ -28,9 +28,11 @@ class ManagerUsersBloc extends DisposeCallbackBaseBloc {
   factory ManagerUsersBloc(final ManagerRepository managerRepository) {
     assert(managerRepository != null);
     final getUsersController = BehaviorSubject<int>();
-    final removeUserController = PublishSubject<User>();
+    final removeUserController =
+        PublishSubject<MapEntry<User, DestroyUserType>>();
     final isLoadingController = BehaviorSubject<bool>.seeded(false);
-    final removingUserIds = BehaviorSubject<List<String>>.seeded([]);
+    final removingUserIds =
+        BehaviorSubject<Map<String, DestroyUserType>>.seeded({});
 
     final controllers = [
       getUsersController,
@@ -40,15 +42,33 @@ class ManagerUsersBloc extends DisposeCallbackBaseBloc {
     ];
 
     final removeUserStream = removeUserController.stream
-        .where((user) => !removingUserIds.value.contains(user.uid))
-        .flatMap((user) => Rx.defer(() async* {
-              removingUserIds.add([...removingUserIds.value, user.uid]);
-              final result = await managerRepository.deleteUser(user);
-              yield result;
-              removingUserIds.add(removingUserIds.value..remove(user.uid));
-            }).doOnError(() =>
-                removingUserIds.add(removingUserIds.value..remove(user.uid))))
-        .share();
+        .where((entry) => !removingUserIds.value.containsKey(entry.key.uid))
+        .flatMap((entry) => Rx.defer(
+              () async* {
+                final user = entry.key;
+                final type = entry.value;
+                User result;
+                if (type == DestroyUserType.REMOVE) {
+                  removingUserIds.add(removingUserIds.value
+                    ..addAll({user.uid: DestroyUserType.REMOVE}));
+                  result = await managerRepository.deleteUser(user);
+                } else if (type == DestroyUserType.BLOCK) {
+                  removingUserIds.add(removingUserIds.value
+                    ..addAll({user.uid: DestroyUserType.BLOCK}));
+                  result = await managerRepository.blockUser(user);
+                } else if (type == DestroyUserType.UNBLOCK) {
+                  removingUserIds.add(removingUserIds.value
+                    ..addAll({user.uid: DestroyUserType.UNBLOCK}));
+                  result = await managerRepository.unblockUser(user);
+                }
+                print(result.toString() + type.toString());
+                yield MapEntry(result, type);
+                removingUserIds.add(removingUserIds.value..remove(user.uid));
+              },
+            ).doOnError(
+              () => removingUserIds
+                  .add(removingUserIds.value..remove(entry.key.uid)),
+            ));
 
     final renderListStream = Rx.merge<ManageUserState>([
       getUsersController.stream
@@ -66,16 +86,19 @@ class ManagerUsersBloc extends DisposeCallbackBaseBloc {
       isLoadingController.stream
           .where((event) => event)
           .map((_) => LoadingUsersState()),
-      removeUserStream.map((_) => DeleteUserSuccess())
-    ]).share();
+      removeUserStream.map((entry) => entry.value == DestroyUserType.REMOVE
+          ? DeleteUserSuccess(idUserDelete: entry.key.uid)
+          : entry.value == DestroyUserType.BLOCK
+              ? BlockUserSuccess(user: entry.key)
+              : UnblockUserSuccess(user: entry.key))
+    ]).publish();
 
-
-    final subscriptions = [renderListStream, removeUserStream];
+    final subscriptions = [renderListStream.connect()];
 
     return ManagerUsersBloc._(
       dispose: DisposeBag([...controllers, ...subscriptions]).dispose,
       loadUsers: getUsersController.add,
-      removeUser: removeUserController.add,
+      destroyUser: removeUserController.add,
       renderListStream$: renderListStream,
       renderItemRemove$: removingUserIds,
     );
